+++ packages/react-components/react-table/src/components/DataGridBody/DataGridBody.types.ts
@@ -7,7 +7,7 @@ export type DataGridBodySlots = TableBodySlots;
 // Use any here since we can't know the user types
 // The user is responsible for narrowing the type downstream
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type RowRenderFunction<TItem = any> = (row: RowState<TItem>) => React.ReactNode;
+export type RowRenderFunction = (row: RowState<any>, ...rest: any[]) => React.ReactNode;
 
 /**
  * DataGridBody Props