+++ packages/react-experiments/webpack.serve.config.js
@@ -1,3 +1,3 @@
-const { resources } = require('@fluentui/scripts/webpack');
+const { resources } = require('@fluentui/scripts-webpack');
 
 module.exports = resources.createLegacyDemoAppConfig();